requirements.txt
aiogram>=3.4,<4.0
aiohttp>=3.9
surrealdb>=0.3
structlog>=24.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
PyMuPDF>=1.24
openpyxl>=3.1
google-generativeai>=0.5
rich>=13.0
